+++ lib/vagrant-lxc/action.rb
@@ -8,6 +8,8 @@ require 'vagrant-lxc/action/destroy'
 require 'vagrant-lxc/action/destroy_confirm'
 require 'vagrant-lxc/action/disconnect'
 require 'vagrant-lxc/action/compress_rootfs'
+require 'vagrant-lxc/action/fetch_ip_with_lxc_attach'
+require 'vagrant-lxc/action/fetch_ip_from_dnsmasq_leases'
 require 'vagrant-lxc/action/forced_halt'
 require 'vagrant-lxc/action/forward_ports'
 require 'vagrant-lxc/action/handle_box_metadata'
@@ -25,7 +27,6 @@ module Vagrant
       # machine back up with the new configuration.
       def self.action_reload
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use CheckDependencies
           b.use Vagrant::Action::Builtin::Call, Created do |env1, b2|
             if !env1[:result]
               b2.use Message, :not_created
@@ -44,10 +45,9 @@ module Vagrant
       # a bootup (i.e. not saved).
       def self.action_boot
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use ClearForwardedPorts
           b.use Vagrant::Action::Builtin::Provision
           b.use Vagrant::Action::Builtin::EnvSet, :port_collision_repair => true
-          # b.use PrepareForwardedPortCollisionParams
+          b.use Vagrant::Action::Builtin::HandleForwardedPortCollisions
           b.use ShareFolders
           b.use Vagrant::Action::Builtin::SetHostname
           b.use ForwardPorts
@@ -58,7 +58,6 @@ module Vagrant
       # This action just runs the provisioners on the machine.
       def self.action_provision
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use CheckDependencies
           b.use Vagrant::Action::Builtin::ConfigValidate
           b.use Vagrant::Action::Builtin::Call, Created do |env1, b2|
             if !env1[:result]
@@ -82,7 +81,6 @@ module Vagrant
       # A precondition of this action is that the container exists.
       def self.action_start
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use CheckDependencies
           b.use Vagrant::Action::Builtin::ConfigValidate
           b.use Vagrant::Action::Builtin::Call, IsRunning do |env, b2|
             # If the VM is running, then our work here is done, exit
@@ -97,7 +95,6 @@ module Vagrant
       # container, configuring metadata, and booting.
       def self.action_up
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use CheckDependencies
           b.use Vagrant::Action::Builtin::ConfigValidate
           b.use Vagrant::Action::Builtin::Call, Created do |env, b2|
             # If the VM is NOT created yet, then do the setup steps
@@ -115,7 +112,6 @@ module Vagrant
       # the virtual machine, gracefully or by force.
       def self.action_halt
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use CheckDependencies
           b.use Vagrant::Action::Builtin::Call, Created do |env, b2|
             if env[:result]
               # TODO: Remove this on / after 0.4
@@ -138,7 +134,6 @@ module Vagrant
       # freeing the resources of the underlying virtual machine.
       def self.action_destroy
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use CheckDependencies
           b.use Vagrant::Action::Builtin::Call, Created do |env1, b2|
             if !env1[:result]
               b2.use Message, :not_created
@@ -162,7 +157,6 @@ module Vagrant
       # This action packages the virtual machine into a single box file.
       def self.action_package
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use CheckDependencies
           b.use Vagrant::Action::Builtin::Call, Created do |env1, b2|
             if !env1[:result]
               b2.use Message, :not_created
@@ -177,10 +171,19 @@ module Vagrant
         end
       end
 
+      # This action is called to read the IP of the container. The IP found
+      # is expected to be put into the `:machine_ip` key.
+      def self.action_fetch_ip
+        Vagrant::Action::Builder.new.tap do |b|
+          b.use Vagrant::Action::Builtin::ConfigValidate
+          b.use FetchIpWithLxcAttach
+          b.use FetchIpFromDnsmasqLeases
+        end
+      end
+
       # This is the action that will exec into an SSH shell.
       def self.action_ssh
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use CheckDependencies
           b.use CheckCreated
           b.use CheckRunning
           b.use Vagrant::Action::Builtin::SSHExec
@@ -190,7 +193,6 @@ module Vagrant
       # This is the action that will run a single SSH command.
       def self.action_ssh_run
         Vagrant::Action::Builder.new.tap do |b|
-          # b.use CheckDependencies
           b.use CheckCreated
           b.use CheckRunning
           b.use Vagrant::Action::Builtin::SSHRun