+++ lib/vagrant-lxc/action/forward_ports.rb
@@ -48,9 +48,9 @@ module Vagrant
                                   message_attributes))
 
             redir_pid = redirect_port(
-              fp[:host_ip]  || "127.0.0.1",
+              fp[:host_ip],
               fp[:host],
-              fp[:guest_ip] || @env[:machine].provider.driver.assigned_ip,
+              fp[:guest_ip] || @env[:machine].provider.ssh_info[:host],
               fp[:guest]
             )
             store_redir_pid(fp[:host], redir_pid)
@@ -72,8 +72,10 @@ module Vagrant
         end
 
         def redirect_port(host_ip, host_port, guest_ip, guest_port)
-          host_ip = "--laddr=#{host_ip}" unless host_ip.empty?
-          redir_cmd = "sudo redir #{host_ip} --lport=#{host_port} --caddr=#{guest_ip} --cport=#{guest_port} 2>/dev/null"
+          params = %W( --lport=#{host_port} --caddr=#{guest_ip} --cport=#{guest_port} )
+          params.unshift "--laddr=#{host_ip}" if host_ip
+          params << '--syslog' if ENV['REDIR_LOG']
+          redir_cmd = "redir #{params.join(' ')} 2>/dev/null"
 
           @logger.debug "Forwarding port with `#{redir_cmd}`"
           spawn redir_cmd
@@ -89,7 +91,7 @@ module Vagrant
         end
 
         def redir_installed?
-          system "sudo which redir > /dev/null"
+          system "which redir > /dev/null"
         end
       end
     end